requirements.txt
# Werkzeug is famous for breaking Flask so pin down a known working version
Werkzeug==2.1.2

# Build dependencies
Flask==2.1.2
Flask-RESTX==0.5.1
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
python-dotenv==0.20.0

# Runtime dependencies
gunicorn==20.1.0
honcho==1.1.0

# Code quality
pylint==2.14.0
flake8==4.0.1

# Testing dependencies
nose==1.3.7
pinocchio==0.4.3
factory-boy==2.12.0
coverage==6.3.2
codecov==2.1.12

# Utilities
httpie==3.2.1
